+++ src/main/java/com/example/airbnbbackend/domain/RoomImage.java
@@ -1,6 +1,7 @@
 package com.example.airbnbbackend.domain;
 
 import jakarta.persistence.*;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
@@ -13,8 +14,10 @@ public class RoomImage {
     @Column(name = "room_image_id")
     private Long id;
 
+    @Getter
     @Column(nullable = false)
     private String imageUrl;
+
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "room_id")
