+++ src/main/java/com/example/airbnbbackend/domain/common/Advantage.java
@@ -1,6 +1,7 @@
 package com.example.airbnbbackend.domain.common;
 
 import jakarta.persistence.*;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.util.ArrayList;
@@ -15,15 +16,19 @@ public class Advantage {
     @Column(name = "advantage_id")
     private Long id;
 
+    @Getter
     @Column(nullable = false)
     private String name;
 
+    @Getter
     @Column(nullable = false)
     private String imageUrl;
 
+    @Getter
     @Column(nullable = false)
     private String description;
 
+    @Getter
     @OneToMany(mappedBy = "advantage", cascade = CascadeType.ALL)
     private List<RoomAdvantage> roomAdvantages = new ArrayList<>();
 }
