+++ src/main/java/com/example/airbnbbackend/domain/common/RoomAdvantage.java
@@ -2,6 +2,7 @@ package com.example.airbnbbackend.domain.common;
 
 import com.example.airbnbbackend.domain.Room;
 import jakarta.persistence.*;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
@@ -14,11 +15,14 @@ public class RoomAdvantage {
     @Column(name = "room_advantage_id")
     private Long id;
 
+
+    @Getter
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "advantage_id")
     private Advantage advantage;
 
-    @OneToOne(fetch = FetchType.LAZY)
+    @Getter
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "room_id")
     private Room room;
 }
