+++ src/main/resources/DB/data.sql
@@ -236,7 +236,7 @@ VALUES ( 1,
         '/images/categoryIcon/categoryIcon5.jpeg',
         '저택'),
        (6,
-        '/categoryIcon/categoryIcon6.jpeg',
+        '/images/categoryIcon/categoryIcon6.jpeg',
         '방'),
        (7,
         '/images/categoryIcon/categoryIcon7.jpeg',
